Move schema definition classes into `mysql/schema_definitions.rb`
Current master branch includes many schema definition improvements in
MySQL. It extract these features to the appropriate file.
